requirements.txt
apache-airflow==3.1.3
boto3==1.42.0
colorlog==6.10.1
dotenv==0.9.9
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.4
pyspark==4.0.1
requests==2.32.5